APIGOV-22993 - move common update instance to util
